+++ Changelog.md
@@ -3,7 +3,8 @@ Lilu Changelog
 
 #### v1.2.4
 - Internalize new APIs from 1.2.3
-- Added new EFI runtime API
+- Added new EFI runtime API with custom variable extensions
+- Added new RTC storage API
 
 #### v1.2.3
 - Added CPU information API for cpu families and generations